Account for additional cases where Run might be stale
We also need to rescue and retry in job-started, as it's possible
for a status change to occur (ie. to pausing or cancelling) from the UI
in between the write in #job_running and the write in #job_started
(notably this is a problem if the Tasks' collection is large, as there
might be a delay in perform @task.count).

We also need to account for scenarios in which an update occuring in
between the read and write in the RunsController causes the run
to be stale. For example, if a tick_count has been updated via the
ticker in between the read and the write.

Co-authored-by: Étienne Barrié <etienne.barrie@shopify.com>