Add integrity hash to Link header for stylesheets
The browser currently warns on preload because it attempts to preload
a Link, but since the integrity hash is missing, it does not use the
Link. Our system tests fail because of this warning.

To fix, we override #stylesheet_link_tag in ApplicationHelper to store
the integrity hash in the Link header.

This can be reverted when a fix is shipped upstream to Rails.

Co-authored-by: Étienne Barrié <etienne.barrie@shopify.com>