Avoid usage of private_constant
This just make hard to use reflection on Ruby classes breaking tools like
tapioca. The value we want to provide with this call is telling people to not
use this class, but that is already achieved by `@api private`.
If people really want to use this constant even if we told them to not do it,
isn't `private_constant` that will stop them.