Postpone reenqueuing the iteration job until after callbacks are done
This PR fixes a bug that occurs when a job gets interrupted and is
reenqueued faster than the original job takes to shut down. This results
in race conditions with the @run and results in errors due to invalid
status transitions (and notably results in a run showing up as interrupted
when it is actually running).

To solve this, we should delay reenqueuing the job until after all callbacks
have completed, which can be done by calling #reenqueue_iteration_job
in a prepended after_perform callback in TaskJob, and skipping the call
that JobIteration performs in #iterate_with_enumerator.