Make Active Storage optional
This allows loading the gem and running regular Tasks when
active_storage/engine is not required in config/application.rb.

Not having required active_storage/engine will have the same result as
having it required but not configured. An error is raised when a CSV
file is uploaded, but no earlier.