Raise early when Active Storage is not setup
Instead of waiting until a CSV file is uploaded, this raises an
exception early when a CSV Task is defined by calling csv_collection and
Active Storage is not loaded or its tables created.

I reused part of the message that Rails emits when a file is attached
and one of the Active Storage tables does not exist.