Re-enable Rubocop documentation rules
Since rubocop-shopify was using DisabledByDefault, simply adding the
rules in our config enabled them, like the documentation explains:

> All cops are then disabled by default. Only cops appearing in user
configuration files are enabled. Enabled: true does not have to be set
for cops in user configuration. They will be enabled anyway.

But now rubocop-shopify moved to marking the specific rules as disabled
instead, which does not enable them automatically when they are used.