Yield the run in Runner.run
When wrapping this gem for Shopify/shopify, we want to get the newly
created or resumed Run from Runner.run.

This yields the current run only if a block is passed to Runner.run,
without documenting that feature since Run is a private API that most
apps shouldn't tinker with.