Handle job still existing for a cancelled run
If a Run is force-cancelled but the job still exists in the queue, when
it gets worked off, it causes an exception because of the state changes
we expect.

The Run moves from cancelled to running, which causes an exception which
is rescued and the Run is moved to errored but that causes another
exception moving from cancelled to errored.

We now consider a cancelled Run as if it is cancelling, and finish the
job early without processing any items.