Don't raise when destroying a run if ActiveStorage isn't installed
ActiveStorage is required by default in new Rails applications.
In `config/application.rb`, we have:

```rb
require "rails/all"
```

This means that `defined?(ActiveStorage)` and
`respond_to?(:has_one_attached)` return true by default. However, it
does not mean that ActiveStorage was _installed_ and that the database
tables were created.

In the Run model, we created the association with the CSV file when
ActiveStorage was _defined_ even if it wasn't _installed_:

```rb
elsif respond_to?(:has_one_attached)
  has_one_attached :csv_file
end
```

When trying to destroy a run, we were getting an error because the
`has_one_attached` method creates callbacks that try to destroy
attachments before we can destroy the run:

```txt
Could not find table 'active_storage_attachments'
(ActiveRecord::StatementInvalid)
```

To fix this, I extracted a method
`MaintenanceTasks.active_storage_installed?` that checks if
ActiveStorage really is installed in the main application.

We can then reuse this method at other places in the code to check if
ActiveStorage is installed.