Rename `Task.available_task` to `.load_all` and deprecate
`Task.available_tasks` not only returns all descendants of the `Task`
class, it also `require`s all files in the application that may be
defining such class.
`available_tasks` does not carry explicit meaning of how heavy the
operation could be, so this renames it to `load_all`, which should be
better at communicating that.