Support Rack 3 by renaming the name attributes of inputs
On the task pages, we would create runs for a task with attributes by
having inputs named `"[task_arguments][some_attribute]"`.

This is because we use `fields_for` in a `FormBuilder` initialized with
a url, i.e. `form_with(url:)`, which doesn't have an `object_name`, so
when nested fields are generated, there's no object name before the
first `[`.

Rack 3 changes the way form data is parsed, which causes all parameters
to be available in `params["[task_arguments]"]` instead of
`params[:task_arguments]`.

Instead of handling this, we use `fields_for` directly on the view, not
on the `FormBuilder` object, and we use the opportunity to rename
`task_arguments` to `task`.

Now the inputs in the task page for a run will be named
`task[some_attribute]` instead of `[task_arguments][some_attribute]`.