Fix timezone test (#930)
The test was created when DST was on, and now we're not, so this causes
a one-hour difference in the utc_offset. We could add some logic around
this, but if we reuse the date and format it instead of using a literal,
this should ensure they're both in DST or not.