Fix statically resolving renders in Task#show
Rails tries to parse `render` calls in partials to generate a tree of
dependencies that should be included in the cache key for a
template/partial. The `RenderParser` is also used by gems such as
ActionView::Precompiler and LooseErbs for similar purposes (identifying
template/partial dependencies).

Currently, `render @task.active_runs` and `render @runs_page.records`
cannot be resolved by the `RenderParser` (this is a documented
shortcoming of the `RenderParser`).

This commit makes the two partial dependencies explicit as recommended
in the `ActionView::CacheHelper` docs. This should not change the
behavior of the renders, but it does enable the `RenderParser` to
identify the partials being rendered.