Add configurable status reload frequency
This allows tasks to specify how often the run status should be reloaded during iteration.

By default, the status is reloaded on every iteration (frequency of 1), but this can be reduced to
improve performance. Note that decreasing the reload frequency impacts how quickly your task will stop
if it is paused or interrupted.

This is especially useful if the task doesn't need to check for cancellation/pausing very often
and can reduce load on the database.