Revert "Avoid incrementing in-memory lock version on tick count updates" (#1326)
* Revert "Avoid incrementing lock version on tick count updates (#1241)"

This reverts commit 1a0efdb9ab5f9781db5edfb2e2ce904f7cdc4f8e.

* Fix test for status reload frequency

The `travel(0.5)` within the block never actually worked properly; we were
only reloading status after the first iteration. The second status reload
call was _actually_ triggered by us hitting a StaleObjectError when calling
`persist_transition` in the `after_perform` callback.

Now, with in-memory lock increments, we don't hit that StaleObjectError
anymore, so we do only want to assert that `#reload_status` is called once.

We also don't need to manually increment the reload status call count; we can
just assert that `#reload_status` is called once in the first place.

* Don't depend on MaintenanceTasks.status_reload_frequency in the test

---------

Co-authored-by: Étienne Barrié <etienne.barrie@shopify.com>