Encode cursors as JSON
Moving forward, encode cursors as JSON for new runs. Existing runs will
continue to use the old cursor encoding mechanism, which was purely
taking the cursor value and coercing it to a string via ActiveRecord.

This is accomplished by tracking how the cursor is encoded on the run
model via a new boolean column, `cursor_is_json`. This value will be
false-y for existing runs.

This change adds support for multi-column cursors, which can occur when:

* A task specifies multiple `cursor_columns`, or
* When iterating over an ActiveRecord collection that has a multi-column
  primary key

Fixes #1226